Overview

This documentation and the code in NEOs is divided into four blocks, each one representing a different aspect of the study of Near-Earth Objects:

The astrometry section is dedicated to fetch, read and write optical and radar observations of Near-Earth objects in a variety of formats.

The propagation section is meant to integrate in time an initial condition using various dynamical models.

The orbit determination section is used to determine the orbit of an asteroid, either from scratch or starting from a pre-existing orbit.

The impact monitoring section deals with the computation of the probability that a given object collides with a planet.

Parameters

NEOs exports a structure called Parameters which contains the most important and repeated arguments of the package's functions. To set the parameters in given session, simply run

using NEOs

# We can use the default values...
params = Parameters()
# or customize some parameters
params = Parameters(params; maxsteps = 1_000, order = 15, abstol = 1E-12)

To see a list of all the available parameters, check the documentation of Parameters.

NEOs.Parameters — Type

Parameters([::Parameters{T};] kwargs...) where {T <: Real}

A collection of the most important parameters in NEOs.jl functions.

Common

verbose::Bool: whether to print output or not (default: true).

Propagation

maxsteps::Int: maximum number of steps for the integration (default: 500).

order::Int: order of Taylor expansions wrt time (default: 25).

abstol::T: absolute tolerance used to compute propagation timestep (default: 1e-20).

parse_eqs::Bool: whether to use the specialized method of jetcoeffs or not (default: true).

bwdoffset/fwdoffset::T: days to propagate beyond first (bwd) / last (fwd) observation (default: 0.5).

coeffstol::T: maximum size of the coefficients (default: 10.0).

The following parameters are only used when propagating the Marsden et al. (1973) nongravitational accelerations model:

marsden_coeffs::NTuple{3, T}: Yarkovsky effect (A2), solar radiation pressure (A1), and normal component (A3) coefficients (default: (0.0, 0.0, 0.0)).

marsden_scalings::NTuple{3, T}: scaling factors for the jet transport perturbation to the coefficients above (default: (0.0, 0.0, 0.0)).

marsden_radial::NTuple{5, T}: radial function constants; i.e. coefficient α, normalizing distance r₀ [au], and exponents m, n and k (default: (1.0, 1.0, 2.0, 0.0, 0.0))

Gauss Method

safegauss::Bool: whether to try Gauss Method only when exactly three tracklets are available (default: true).

refscale::Symbol: horizontal scale for ADAM refinement of Gauss preliminary orbit (default: :log).

gaussorder::Int: order of the jet transport perturbation (default: 6).

Minimization over the MOV

H_max::T: maximum absolute magnitude (default: 34.5).

a_max::T: maximum semimajor axis (default: 100.0).

adamiter::Int: maximum number of iterations for ADAM optimizer (default: 200).

adammode::Bool: whether to perform ADAM iterations with all the observations (default: true).

adamQtol::T: target function relative tolerance (default: 0.001).

mmovproject::Bool: whether to project the orbits onto the admissible region (default: true).

tsaorder::Int: order of the jet transport perturbation (default: 6).

Least Squares

lsiter::Int: maximum number of iterations for leastsquares (default: 5).

jtlsiter::Int: maximum number of iterations for jtls (default: 5).

jtlsorder::Int: order of the jet transport perturbation in jtls (default: 6).

significance::T: chi-square significance level (default: 0.99).

jtlsmask::Bool: whether to use isjtlsfit to skip bad-conditioned preliminary orbits in jtls (default: true).

jtlsproject::Bool: whether to project the orbits onto the admissible region (default: false).

Outlier Rejection

outrej::Bool: whether to perform outlier rejection during least squares iterations (default: false).

χ2_rec::T: recovery threshold (default: 7.0).

χ2_rej::T: rejection threshold (default: 8.0).

fudge::T: rejection fudge term coefficient (default: 400.0).

max_per::T: maximum allowed rejection percentage (default: 10.0).

Physical properties

slope::T: slope parameter (default: 0.15).

albedo::T: albedo (default: 0.14).

density::T: density in kg/m³ (default: 2_600.0).

Parameters

Across the documentation, the green boxes will indicate which parameters are relevant for a particular block of code.